requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-docx==1.1.0
pydantic==2.6.1
openai==1.12.0
python-dotenv==1.0.1
PyPDF2==3.0.1
fpdf==1.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 